fix: resolve react-doctor warnings from PR #348 review
- Add aria-label to ConflictReviewDialogDiffViewer nav buttons
- Extract MOCK_CONFLICTS to ConflictReviewDialog.mock.ts
- Replace transition-all with transition-[width] in GitLabSyncDialogProgress
- Group sidebar booleans into SidebarState object in WriteModeView
